requirements.txt
django-extensions   #i.e. shell_plus
django-isbn-field
pillow
